requirements.txt
fastapi
uvicorn
torch
numpy
rasterio
pillow